(labs 5-7) refactor: clean up code

+++ lab-5/C++/task-1/Remote.cpp
@@ -27,6 +27,6 @@ void Remote::decVolume() {
     tv->decVolume();
 }
 
-void Remote::goToChannel(int channel) {
+void Remote::goToChannel(short channel) {
     tv->setChannel(channel);
 }
